#64 Fix WebSocket && Added Comment
